# api-docs/openapi.yaml
openapi: "3.0.3"
info:
  title: 珠宝通 ERP API
  description: |
    珠宝通零售连锁门店主动式管理系统 — 前后端共用接口文档。

    ## 使用说明
    - **后端**：以此文档为契约，确保 Controller 实现与文档一致
    - **前端**：以此文档为参考，编写 API 调用代码；也可用 openapi-generator 自动生成 TypeScript 客户端

    ## 认证
    除登录/注册外，所有接口需在请求头携带 JWT Token：
    ```
    Authorization: Bearer <token>
    ```

    ## 统一响应格式
    - 成功：`{ "code": 200, "msg": "success", "data": {...} }`
    - 失败：`{ "code": 500, "msg": "错误描述" }`
    - 参数错误：`{ "code": 400, "msg": "参数错误描述" }`
    - 未登录：`{ "code": 401, "msg": "请先登录" }`
    - 无权限：`{ "code": 403, "msg": "权限不足" }`

    ## 分页规范
    - 请求：`{ "page": 1, "pageSize": 20 }`（Query 参数）
    - 响应：`{ "list": [...], "total": 100, "page": 1, "size": 20 }`
  version: "1.0.0"
  contact:
    name: 珠宝通开发团队
servers:
  - url: http://localhost:8080/api
    description: 本地开发环境
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResult:
      type: object
      properties:
        code:
          type: integer
          description: 状态码，200=成功
          example: 200
        msg:
          type: string
          description: 消息
          example: success
        data:
          description: 响应数据（类型视具体接口而定）

    # ---- 分页 ----
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: 用户名
          example: "admin"
        password:
          type: string
          format: password
          description: 密码
          example: "123456"
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: 用户名
        password:
          type: string
          format: password
          description: 密码
        phone:
          type: string
          description: 手机号
    LoginResult:
      type: object
      properties:
        token:
          type: string
          description: JWT Token
        userId:
          type: integer
          format: int64
        username:
          type: string
        realName:
          type: string
          description: 真实姓名
        avatar:
          type: string
          description: 头像 URL
        phone:
          type: string
        email:
          type: string
        storeId:
          type: integer
          format: int64
        storeName:
          type: string
          description: 所属门店名称
        roles:
          type: array
          items:
            type: string
          description: 角色编码列表
        roleNames:
          type: array
          items:
            type: string
          description: 角色名称列表
        permissions:
          type: array
          items:
            type: string
          description: 权限编码列表
    ProfileUpdateRequest:
      type: object
      properties:
        phone:
          type: string
        email:
          type: string
        avatar:
          type: string
        timezone:
          type: string
        language:
          type: string
        dateFormat:
          type: string
        notifySystem:
          type: boolean
        notifyOrder:
          type: boolean
        notifyInventory:
          type: boolean
        notifyMarketing:
          type: boolean
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword]
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          format: password

    # ---- 商品 ----
...
security:
  - BearerAuth: []

# ============================================================
# API 路径
# ============================================================
paths:
...
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'
  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册（默认导购角色）
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功，直接返回登录信息
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResult'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResult'
...
  /auth/profile:
    put:
      tags: [认证]
      summary: 更新个人信息
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileUpdateRequest'
      responses:
        '200':
          description: 更新成功
  /auth/password:
    put:
      tags: [认证]
      summary: 修改密码
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功
...